Add INSUPERCALL flag to pickle format.
Reason: It is really hard to establish when a context is a
inSuperCall context when unpickling trees. The unpickler does
not have the ability to look inside (untyped) trees, so  it is
tricky/messy to detect a super.<init> or this.<init> Apply in a
constructor body.

Also: Always pickle template constructors.