Halfway to yet another scheme for handling positions
The previous attempt was very fragile, and did not play
well with lazy trees, despite best intentions. It did
not work correctly for normal trees, and it seemed anyeway
very hard to extend this to annotations.

The new scheme, as outlined in PicklerFormat, is
simpler and naturally handles lazy trees and annotations.