New phase to check for accidental captures in closures
`CheckCaptures` is an experimental phase that will check the captured
environment of every Closure to make sure that it doesn't contain any
type marked `@checkCaptures`, unless the expected type of the closure is
marked `@allowCaptures`. This is used to check that we do not
accidentally capture `Context` which can lead to various issues,
especially in interactive uses of the compiler since using a `Context`
from a previous run is very likely to throw a StaleSymbolError.

This phase will not be merged in dotty since it requires a lot of manual
annotations to be useful (more than 150), but it has already lead to
finding several bugs:
- https://github.com/lampepfl/dotty/pull/1974

See also https://github.com/TiarkRompf/scala-escape for a much more
advanced and safer alternative that could be an inspiration for future
work.