typedUnApply#trySelectUnapply: small cleanup
Previously, `specificProto` was a def even though it is always called,
this is because in cece884812143c6c8090ce08c6321bd4a1d52ea6, the usages
of `specificProto` and `genericProto` were swapped. We fix this by only
defining the protos where they are used. Incidentally, this mean that
the calls to UnapplyFunProto will use the correct Context inside
`tryEither`, although in this case this shouldn't matter.