Fix position of class parents, implicit conversions and eta-expansion
All these fixes are mean to make `NavigateAST#pathTo` return the correct
tree for a given position, when multiple trees cover the same position,
the first tree found in depth first order is returned by
`NavigateAST#pathTo` so we need to be careful about the positions we
assign to trees when desugaring:

- The position of class parents was the position of the whole class, now
it's just the position of the class name.

- Implicit conversions had the same position as the expression to be
converted, now they get the start position of the expression.

- When eta-expanding: `foo` to `(x, y) => foo(x, y)` the synthetic
parameters and the arguments got the same position as `foo`, now  they
get the start position of `foo`.