Remove TypeErasure#erasedLub bias for classes
This is removing a disparity between lubs obtained before and after erasure.

Consider this example:

```
if b ast.Trees.Match(???)
else ast.Trees.Literal(???)
```

Before erasure this would be an or type that's later widen to
`ast.Trees.TermTree`. Because of `erasedLub` bias for classes, recomputing
this lub after erasure lead to `ast.Trees.Tree` instead.

This disparity prevents a local optimisation from being Ycheckable;
removing the bias fixes that.