Make TypedTreeCopier be more aggressive.
And create a TimeTravellingCopier that is more correct.
The main difference is that TypedTreeCopier assumes that it does not
need to re-widen or re-lub types that we're widened or lubbed
when previous tree was constructed.

This assumption is false old tree was constructed in a different phase.
TimeTravellingCopier has less aggressive but more slow implementation.