Don't interpolate or simplify overloaded references
Simplifying an overloaded reference risks changing the prefix, which will
lead into trouble using the new overloading scheme (because the only info
is in the denotation, which will be thrown away when the prefix is changed).

Also, both interpolating undetermined variables and simplifying are likely to be
inefficient on overloaded types because they tend to be large. Delaying these operations
means we only have to perform them on alternatives that might plausibly match the
expected type.