Fail on all stale references by default
Update stale references only if -Yupdate-state is set, fail with a StaleSymbol
error otherwise.

This required some fixes to the way ScalaShadowing was treated. If we compile a
ScalaShadowing (as is done in tasty-bootstrap) we need to also overwrite the
mirrored symbol in the scala package. Otherwise we would get a stale reference
when looking up the latter.