Make sure bringForward returns a denotation valid in current period
There was a scenario in the IDE where this was not the case: If the owner of
a denotation was deleted in the new run, the old symbol gets an updated validity.
Then, a call to bringForward of a member of that invalid owner will lookup
the symbol in its owner, yielding itself. But the validity still was the previous period.